RIP local module search support code
