+++ website/src/types/state.ts
@@ -1,6 +1,5 @@
 import {
   AppState,
-  ModuleFinderState,
   PlannerState,
   Requests,
   SettingsState,
@@ -19,7 +18,6 @@ export type State = {
   app: AppState;
   theme: ThemeState;
   settings: SettingsState;
-  moduleFinder: ModuleFinderState;
   planner: PlannerState;
   undoHistory: UndoHistoryState;
 };
