+++ website/src/views/modules/ModuleFinderContainer.tsx
@@ -1,378 +1,202 @@
-import * as React from 'react';
-import { RouteComponentProps, withRouter } from 'react-router-dom';
-import { connect } from 'react-redux';
-import axios from 'axios';
-import produce from 'immer';
-import { each, mapValues, values } from 'lodash';
+import React, { useState } from 'react';
+import {
+  ReactiveBase,
+  ReactiveList,
+  SelectedFilters,
+  MultiList,
+  MultiDataList,
+  MultiDropdownList,
+  MultiRange,
+} from '@appbaseio/reactivesearch';
+import { map } from 'lodash';
 
-import { ModuleInformation } from 'types/modules';
-import { AnyGroup, FacultyDepartments, FilterGroups, PageRange, PageRangeDiff } from 'types/views';
+import { attributeDescription } from 'types/modules';
 
+import ModuleFinderItem from 'views/components/ModuleFinderItem';
 import ModuleFinderList from 'views/modules/ModuleFinderList';
 import ModuleSearchBox from 'views/modules/ModuleSearchBox';
 import ChecklistFilters from 'views/components/filters/ChecklistFilters';
 import DropdownListFilters from 'views/components/filters/DropdownListFilters';
 import ApiError from 'views/errors/ApiError';
+import Warning from 'views/errors/Warning';
 import LoadingSpinner from 'views/components/LoadingSpinner';
 import SideMenu, { OPEN_MENU_LABEL } from 'views/components/SideMenu';
 import { Filter } from 'views/components/icons';
 import Title from 'views/components/Title';
 import Omelette, { matchEgg } from 'views/components/Omelette';
+
+import config from 'config';
 import { forceInstantSearch } from 'utils/debug';
-import {
-  defaultGroups,
-  DEPARTMENT,
-  EXAMS,
-  FACULTY,
-  LEVELS,
-  MODULE_CREDITS,
-  SEMESTER,
-  serializeGroups,
-  updateGroups,
-  ATTRIBUTES,
-} from 'utils/moduleFilters';
-import { createSearchFilter, SEARCH_QUERY_KEY, sortModules } from 'utils/moduleSearch';
-import nusmods from 'apis/nusmods';
-import { resetModuleFinder } from 'actions/moduleFinder';
-import FilterGroup from 'utils/filters/FilterGroup';
-import HistoryDebouncer from 'utils/HistoryDebouncer';
-import { defer } from 'utils/react';
-import { breakpointUp, queryMatch } from 'utils/css';
 import { captureException } from 'utils/error';
-import { State as StoreState } from 'types/state';
 import styles from './ModuleFinderContainer.scss';
 
-export type Props = {
-  searchTerm: string;
-  resetModuleFinder: () => unknown;
-} & RouteComponentProps;
-
-export type State = {
-  loading: boolean;
-  page: PageRange;
-  modules: ModuleInformation[];
-  filterGroups: FilterGroups;
-  isMenuOpen: boolean;
-  error: Error | null;
-};
-
-// Threshold to enable instant search based on the amount of time it takes to
-// run the filters on initial render. This is only an estimate since only
-// ~50% of the time is spent on filters (the other 50% on rendering), and we can
-// only benchmark filtering, not rendering, so we err on using a lower threshold
-const INSTANT_SEARCH_THRESHOLD = 300;
+export type Props = {};
 
 const pageHead = <Title>Modules</Title>;
 
-export function mergePageRange(prev: PageRange, diff: PageRangeDiff): PageRange {
-  const next = { ...prev };
+const reactiveBaseResetTheme = {
+  typography: {
+    fontFamily: 'inherit',
+    fontSize: 'inherit',
+  },
+};
 
-  // Current page is SET from the diff object
-  if (diff.current != null) {
-    next.current = diff.current;
-  }
-
-  // Start and pages are ADDED from the diff object
-  if (diff.start != null) next.start += diff.start;
-  if (diff.loaded != null) next.loaded += diff.loaded;
-
-  return next;
-}
-
-export class ModuleFinderContainerComponent extends React.Component<Props, State> {
-  history: HistoryDebouncer;
-
-  unlisten: () => void;
-
-  state: State;
-
-  useInstantSearch = false;
-
-  constructor(props: Props) {
-    super(props);
-
-    // Set up history debouncer and history listener
-    this.history = new HistoryDebouncer(props.history);
-    this.unlisten = props.history.listen((location) => this.onQueryStringChange(location.search));
-
-    this.state = {
-      filterGroups: {},
-      page: this.startingPageRange(),
-      loading: true,
-      modules: [],
-      isMenuOpen: false,
-      error: null,
-    };
-  }
-
-  componentDidMount() {
-    this.loadPageData();
-  }
-
-  componentDidUpdate(prevProps: Props) {
-    if (this.props.searchTerm !== prevProps.searchTerm) {
-      this.onSearch(this.props.searchTerm);
-    }
-  }
-
-  componentWillUnmount() {
-    this.props.resetModuleFinder();
-    this.unlisten();
-  }
-
-  // Event handlers
-  onQueryStringChange(query: string) {
-    const { filterGroups } = this.state;
-
-    // Trim the starting '?' character
-    if (query.replace(/^\?/, '') !== serializeGroups(filterGroups)) {
-      this.setState({
-        filterGroups: updateGroups(filterGroups, query),
-      });
-    }
-  }
-
-  onFilterChange = (newGroup: AnyGroup, resetScroll: boolean = true) => {
-    this.setState(
-      produce<State>((draft) => {
-        // Update filter group with its new state
-        draft.filterGroups[newGroup.id] = newGroup;
-
-        // Reset back to the first page
-        draft.page.start = 0;
-        draft.page.current = 0;
-      }),
-
-      () => {
-        this.updateQueryString();
-
-        // Scroll back to the top
-        if (resetScroll) window.scrollTo(0, 0);
-      },
-    );
-  };
-
-  onPageChange = (diff: PageRangeDiff) => {
-    this.setState(
-      (prevState) => ({
-        page: mergePageRange(prevState.page, diff),
-      }),
-      this.updatePageHash,
-    );
-  };
-
-  onClearFilter = () => {
-    const filterGroups = mapValues(this.state.filterGroups, (group: FilterGroup<any>) => {
-      // Don't clear search query
-      if (group.id === SEARCH_QUERY_KEY) return group;
-      return group.reset();
-    });
-
-    this.setState({ filterGroups }, this.updateQueryString);
-  };
-
-  onSearch(searchTerm: string) {
-    const filter = createSearchFilter(searchTerm).initFilters(this.state.modules);
-
-    defer(() => {
-      this.onFilterChange(filter, false);
-      this.toggleMenu(false);
-    });
-  }
-
-  loadPageData = () => {
-    this.setState({ error: null });
-
-    // Load module data
-    const modulesRequest = axios
-      .get<ModuleInformation[]>(nusmods.modulesUrl())
-      .then(({ data }) => data);
-
-    // Load faculty-department mapping
-    const facultiesRequest = axios
-      .get<FacultyDepartments>(nusmods.facultyDepartmentsUrl())
-      .then(({ data }) => data);
-
-    // Finally initialize everything
-    Promise.all([modulesRequest, facultiesRequest])
-      .then(([modules, faculties]) => {
-        const filterGroups = defaultGroups(faculties, this.props.location.search);
-
-        // Benchmark the amount of time taken to run the filters to determine if we can
-        // use instant search
-        const start = window.performance && performance.now();
-        each(filterGroups, (group) => group.initFilters(modules));
-        const time = window.performance
-          ? performance.now() - start
-          : // If the user's browser doesn't support performance.now, we assume it is too old for instant search
-            Number.MAX_VALUE;
-
-        const force = forceInstantSearch();
-        if (typeof force === 'boolean') {
-          // Manual override - use 'instant=1' to force instant search, and
-          // 'instant=0' to force non-instant search
-          this.useInstantSearch = force;
-        } else {
-          // By default, only turn on instant search for desktop and if the
-          // benchmark earlier is fast enough
-          this.useInstantSearch =
-            queryMatch(breakpointUp('sm')).matches && time < INSTANT_SEARCH_THRESHOLD;
-        }
-
-        console.info(`${time}ms taken to init filters`); // eslint-disable-line
-        console.info(this.useInstantSearch ? 'Instant search on' : 'Instant search off'); // eslint-disable-line
-
-        this.setState({
-          filterGroups,
-          modules,
-          loading: false,
-        });
-      })
-      .catch((error) => {
-        captureException(error);
-        this.setState({ error });
-      });
-  };
-
-  updateQueryString = () => {
-    // Update query string after state is updated
-    this.history.push({
-      ...this.props.history.location,
-      search: serializeGroups(this.state.filterGroups),
-    });
-  };
-
-  updatePageHash = () => {
-    // Update the location hash so that users can share the URL and go back to the
-    // correct page when the going back in history
-    const { current } = this.state.page;
-    this.history.push({
-      ...this.props.history.location,
-      hash: current ? `page=${current}` : '',
-    });
-  };
-
-  toggleMenu = (isMenuOpen: boolean) => this.setState({ isMenuOpen });
-
-  filterGroups(): AnyGroup[] {
-    return values(this.state.filterGroups);
-  }
-
-  startingPageRange(): PageRange {
-    const hashMatch = this.props.location.hash.match(/page=(\d+)/);
-    const start = hashMatch ? parseInt(hashMatch[1], 10) : 0;
-
-    return {
-      start,
-      current: start,
-      loaded: 1,
-    };
-  }
-
-  render() {
-    const { filterGroups: groups, isMenuOpen, modules, loading, page, error } = this.state;
-
-    if (error) {
-      return <ApiError dataName="module information" retry={this.loadPageData} />;
-    }
-
-    if (loading) {
-      return (
-        <div>
-          {pageHead}
-          <LoadingSpinner />
-        </div>
-      );
-    }
-
-    // Set up filter groups and sort by relevance if search is active
-    let filteredModules = FilterGroup.apply(modules, this.filterGroups());
-    if (this.props.searchTerm) {
-      filteredModules = sortModules(this.props.searchTerm, filteredModules);
-    }
-
-    const filterProps = {
-      groups: values(groups),
-      onFilterChange: this.onFilterChange,
-    };
-
-    // The clear filter button is only shown when there is at least one filter
-    // active
-    const showClearFilters = this.filterGroups().some(
-      (group) => group.isActive() && group.id !== SEARCH_QUERY_KEY,
-    );
-
-    return (
-      <div className="modules-page-container page-container">
-        {pageHead}
-
+const ModuleFinderContainer: React.FunctionComponent<Props> = () => {
+  const [isMenuOpen, setMenuOpen] = useState(false);
+  return (
+    <div className="modules-page-container page-container">
+      {pageHead}
+      <ReactiveBase
+        className={styles.reactiveBase}
+        app="modules"
+        url="http://localhost:9200"
+        theme={reactiveBaseResetTheme}
+      >
         <div className="row">
           <div className="col">
             <h1 className="sr-only">Module Finder</h1>
 
-            <ModuleSearchBox useInstantSearch={this.useInstantSearch} />
+            <ModuleSearchBox
+              componentId="q"
+              showFilter={false}
+              URLParams
+              dataField={['moduleCode^10', 'title^3', 'description']}
+            />
 
-            {matchEgg(this.props.searchTerm) ? (
-              <Omelette query={this.props.searchTerm} />
-            ) : (
-              <ModuleFinderList
-                modules={filteredModules}
-                page={page}
-                onPageChange={this.onPageChange}
+            <ul className="modules-list">
+              <ReactiveList
+                componentId="page"
+                dataField="moduleCode"
+                pagination
+                URLParams
+                react={{
+                  and: ['q', 'sem', 'level', 'mcs', 'faculty', 'department', 'attrs'],
+                }}
+                innerClass={{
+                  pagination: styles.pagination,
+                }}
+                loader={<LoadingSpinner />}
+                renderItem={(res) => <ModuleFinderItem key={res.moduleCode} module={res} />}
+                renderResultStats={(res) => {
+                  let { numberOfResults } = res;
+                  // numberOfResults is an object instead of a number in Elasticsearch 7.
+                  // See: https://github.com/appbaseio/reactivesearch/issues/948#issuecomment-488661618
+                  if (!Number.isInteger(numberOfResults)) {
+                    numberOfResults = numberOfResults.value;
+                  }
+                  return <div className="module-page-divider">{numberOfResults} modules found</div>;
+                }}
+                renderNoResults={() => <Warning message="No modules found" />}
               />
-            )}
+            </ul>
           </div>
 
           <div className="col-md-4 col-lg-3">
             <SideMenu
               isOpen={isMenuOpen}
-              toggleMenu={this.toggleMenu}
+              toggleMenu={() => setMenuOpen(!isMenuOpen)}
               openIcon={<Filter aria-label={OPEN_MENU_LABEL} />}
             >
               <div className={styles.moduleFilters}>
                 <header className={styles.filterHeader}>
                   <h3>Refine by</h3>
-                  {showClearFilters && (
-                    <button
-                      className="btn btn-link btn-sm"
-                      type="button"
-                      onClick={this.onClearFilter}
-                    >
-                      Clear Filters
-                    </button>
-                  )}
                 </header>
 
-                <ChecklistFilters group={groups[SEMESTER]} {...filterProps} />
+                <SelectedFilters showClearAll clearAllLabel="Clear filters" />
 
-                <ChecklistFilters group={groups[LEVELS]} {...filterProps} />
-
-                <ChecklistFilters group={groups[EXAMS]} {...filterProps} />
-
-                <ChecklistFilters group={groups[MODULE_CREDITS]} {...filterProps} />
-
-                <DropdownListFilters group={groups[FACULTY]} {...filterProps} />
-
-                <DropdownListFilters group={groups[DEPARTMENT]} {...filterProps} />
-
-                <ChecklistFilters group={groups[ATTRIBUTES]} {...filterProps} />
+                <MultiDataList
+                  componentId="sem"
+                  dataField="semesterData.semester"
+                  title="Offered In"
+                  filterLabel="Offered In"
+                  URLParams
+                  data={map(config.semesterNames, (semName, semNumStr) => ({
+                    label: semName,
+                    value: semNumStr,
+                  }))}
+                  defaultValue={Object.values(config.semesterNames)}
+                  react={{
+                    and: ['q', 'level', 'mcs', 'faculty', 'department', 'attrs'],
+                  }}
+                  showCount
+                  showSearch={false}
+                />
+                <MultiList
+                  componentId="level"
+                  dataField="moduleCode.level"
+                  title="Level"
+                  filterLabel="Level"
+                  sortBy="asc"
+                  showSearch={false}
+                  URLParams
+                  react={{
+                    and: ['q', 'sem', 'mcs', 'faculty', 'department', 'attrs'],
+                  }}
+                />
+                <MultiRange
+                  componentId="mcs"
+                  dataField="moduleCredit"
+                  title="Module Credit"
+                  filterLabel="MCs"
+                  URLParams
+                  data={[
+                    { start: 0, end: 3, label: '0-3 MC' },
+                    { start: 4, end: 4, label: '4 MC' },
+                    { start: 5, end: 8, label: '5-8 MC' },
+                    { start: 8, end: 300, label: 'More than 8 MC' },
+                  ]}
+                  react={{
+                    and: ['q', 'sem', 'level', 'faculty', 'department', 'attrs'],
+                  }}
+                />
+                <MultiDropdownList
+                  componentId="faculty"
+                  dataField="faculty.keyword"
+                  title="Faculty"
+                  filterLabel="Faculty"
+                  queryFormat="or"
+                  showCheckbox
+                  URLParams
+                  react={{
+                    and: ['q', 'sem', 'level', 'mcs', 'department', 'attrs'],
+                  }}
+                />
+                <MultiDropdownList
+                  componentId="department"
+                  dataField="department.keyword"
+                  title="Department"
+                  filterLabel="Department"
+                  queryFormat="or"
+                  showCheckbox
+                  URLParams
+                  react={{
+                    and: ['q', 'sem', 'level', 'mcs', 'faculty', 'attrs'],
+                  }}
+                />
+                <MultiDataList
+                  componentId="attrs"
+                  dataField="attributes"
+                  title="Other attributes"
+                  URLParams
+                  data={['su', 'grsu', 'ssgf', 'sfs', 'lab', 'ism'].map((attr) => ({
+                    label: attributeDescription[attr] || attr,
+                    value: attr,
+                  }))}
+                  multiSelect
+                  showSearch={false}
+                  customQuery={(values) => {
+                    if (!Array.isArray(values) || values.length === 0) return null;
+                    const should = values.map((v) => ({ match: { [`attributes.${v}`]: true } }));
+                    return { query: { bool: { should } } };
+                  }}
+                />
               </div>
             </SideMenu>
           </div>
         </div>
-      </div>
-    );
-  }
-}
+      </ReactiveBase>
+    </div>
+  );
+};
 
-const mapStateToProps = (state: StoreState) => ({
-  searchTerm: state.moduleFinder.search.term,
-});
-
-// Explicitly naming the components to make HMR work
-const ModuleFinderContainerWithRouter = withRouter(ModuleFinderContainerComponent);
-const ModuleFinderContainer = connect(
-  mapStateToProps,
-  { resetModuleFinder },
-)(ModuleFinderContainerWithRouter);
 export default ModuleFinderContainer;
